requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
PyYAML>=6.0
pillow>=9.0.0
scipy>=1.7.0
matplotlib>=3.4.0

# Machine learning and deep learning
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.18.0
librosa>=0.8.1
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Audio processing
pydub>=0.25.1
soundfile>=0.10.3

# Video processing
opencv-python>=4.5.5
imageio>=2.16.0
decord>=0.6.0

# Interface
flask>=2.0.0
waitress>=2.1.1
flask-cors>=3.0.10

# Reporting and visualization
jinja2>=3.0.0
pdfkit>=1.0.0

# Testing and development
pytest>=7.0.0
black>=22.1.0
isort>=5.10.0
flake8>=4.0.0